Make more tiles walkable
The tiles with rocks just on the top edge look like they should be
walkable-on, so now they are: I adjusted their collision polygon so
that only the bit with rocks on it is solid. This prevents entering
or leaving them via the solid edge, but they can otherwise be walked
on because our player & enemy are treated as a point in the middle of
the cell.

Then we need to treat them as not-solid for the A* search. To do
this, add a custom data layer to the TileSet which contains a boolean
'navigable' property. Set this property to true on the 4 appropriate
tiles.

Finally I redrew the map so it is a little more interesting.