Make enemy bump into player
To do this we have to turn the player back into a physical body
(CharacterBody2D) and reattach its collision shape. The shape is set to
16×16px even though the sprite extends slightly outside.

We then need to ensure the player has finished its move before the
enemy starts to move. Previously the two turns would actually take place
simultaneously. This allowed the enemy to move through the player in
some situation that I didn't debug. Instead, add a signal to the player
& enemy emitted when the movement is finished.

Ensure the enemy always plays an animation even if they don't move
anywhere.